Fix speeding up and slowing down particles

+++ include/visualizer/particle_box.h
@@ -58,7 +58,7 @@ class ParticleBox {
    * Speeds up or slows down all particles on the screen
    * @param speed_factor - how much the all the particles speed increases by
    */
-  void IncreaseDecreaseSpeed(size_t speed_factor);
+  void IncreaseDecreaseSpeed(float speed_factor);
 
  private:
